plugins/com.google.cloud.tools.eclipse.appengine.localserver.test/src/com/google/cloud/tools/eclipse/appengine/localserver/ServletClasspathProviderTest.java
package com.google.cloud.tools.eclipse.appengine.localserver;

import org.eclipse.jdt.core.IClasspathEntry;
import org.junit.Assert;
import org.junit.Test;

// Must Run as JUnit Plug-in Test in Eclipse.
public class ServletClasspathProviderTest {

  private ServletClasspathProvider provider = new ServletClasspathProvider();
  
  @Test
  public void testResolveClasspathContainer() {
     IClasspathEntry[] result = provider.resolveClasspathContainer(null, null);
     if (result.length == 2) {
       Assert.assertTrue(result[0].getPath().toString().endsWith("servlet-api.jar"));
       Assert.assertTrue(result[1].getPath().toString().endsWith("jsp-api.jar"));
     } else {
       // hack because gcloud is not yet installed on Travis.
       // see https://github.com/GoogleCloudPlatform/gcloud-eclipse-tools/issues/267
       // TODO fix this!
       Assert.assertEquals(0, result.length);
     }
  }

}
